rewrite process init packet
