+++ src/include/networking/network_manager.h
@@ -1,234 +1,134 @@
-//===----------------------------------------------------------------------===//
 //
-//                         Peloton
+// Created by tim on 25/07/17.
 //
-// packet_manager.h
-//
-// Identification: src/include/networking/packet_manager.h
-//
-// Copyright (c) 2015-16, Carnegie Mellon University Database Group
-//
-//===----------------------------------------------------------------------===//
-
-#pragma once
-
-#include <boost/assign/list_of.hpp>
-#include <iostream>
-#include <mutex>
-#include <string>
-#include <unordered_map>
 #include <vector>
 
-#include "common/cache.h"
-#include "common/portal.h"
-#include "common/statement.h"
-#include "tcop/tcop.h"
-#include "networking/marshal.h"
+#include "networking/network_thread.h"
+#include "networking/protocol_handler.h"
 #include "networking/packet_manager.h"
-
-// Packet content macros
-#define NULL_CONTENT_SIZE -1
+#pragma once
 
 namespace peloton {
-
 namespace networking {
+// Buffers used to batch messages at the socket
+struct Buffer {
+  size_t buf_ptr;        // buffer cursor
+  size_t buf_size;       // buffer size
+  size_t buf_flush_ptr;  // buffer cursor for write
+  ByteBuf buf;
 
-typedef std::vector<std::unique_ptr<OutputPacket>> ResponseBuffer;
+  inline Buffer() : buf_ptr(0), buf_size(0), buf_flush_ptr(0) {
+    // capacity of the buffer
+    buf.reserve(SOCKET_BUFFER_SIZE);
+  }
 
+  inline void Reset() {
+    buf_ptr = 0;
+    buf_size = 0;
+    buf_flush_ptr = 0;
+  }
+
+  // single buffer element accessor
+  inline uchar GetByte(size_t &index) { return buf[index]; }
+
+  // Get pointer to index location
+  inline uchar *GetPtr(size_t index) { return &buf[index]; }
+
+  inline ByteBuf::const_iterator Begin() { return std::begin(buf); }
+
+  inline ByteBuf::const_iterator End() { return std::end(buf); }
+
+  inline size_t GetMaxSize() { return buf.size(); }
+};
+/*
+ * NetworkManager - Wrapper for managing socket.
+ * 	B is the STL container type used as the protocol's buffer.
+ */
 class NetworkManager {
  public:
-  NetworkManager();
+  int thread_id;
+  int sock_fd;                    // socket file descriptor
+  struct event *event = nullptr;  // libevent handle
+  short event_flags;              // event flags mask
 
-  ~NetworkManager();
+  SSL* conn_SSL_context = nullptr;          // SSL context for the connection
 
-  /* Routine to deal with the first packet from the client */
-  int ProcessInitialPacket(InputPacket* pkt);
-
-  /* Routine to deal with general Startup message */
-  bool ProcessStartupPacket(InputPacket* pkt, int32_t proto_version);
-
-  /* Routine to deal with SSL request message */
-  bool ProcessSSLRequestPacket(InputPacket *pkt);
-
-  /* Main switch case wrapper to process every packet apart from the startup
-   * packet. Avoid flushing the response for extended protocols. */
-  bool ProcessPacket(InputPacket* pkt, const size_t thread_id);
-
-  /* Manage the startup packet */
-  //  bool ManageStartupPacket();
-  void Reset();
-
-  // Returns a vector of all the PreparedStatements that this NetworkManager has
-  // that reference the given table id
-  const std::vector<Statement*> GetPreparedStatements(oid_t table_id) {
-    return table_statement_cache_[table_id];
-  }
-
-  void InvalidatePreparedStatements(oid_t table_id);
-
-  // Ugh... this should not be here but we have no choice...
-  void ReplanPreparedStatement(Statement* statement);
-
-  // Check existence of statement in cache by name
-  // Return true if exists
-  bool ExistCachedStatement(std::string statement_name) {
-    auto statement_cache_itr = statement_cache_.find(statement_name);
-    return statement_cache_itr != statement_cache_.end();
-  }
-
-  //===--------------------------------------------------------------------===//
-  // STATIC HELPERS
-  //===--------------------------------------------------------------------===//
-
-  // Deserialize the parameter types from packet
-  static size_t ReadParamType(InputPacket* pkt, int num_params,
-                              std::vector<int32_t>& param_types);
-
-  // Deserialize the parameter format from packet
-  static size_t ReadParamFormat(InputPacket* pkt, int num_params_format,
-                                std::vector<int16_t>& formats);
-
-  // Deserialize the parameter value from packet
-  static size_t ReadParamValue(
-      InputPacket* pkt, int num_params, std::vector<int32_t>& param_types,
-      std::vector<std::pair<int, std::string>>& bind_parameters,
-      std::vector<type::Value>& param_values, std::vector<int16_t>& formats);
-
-  static std::vector<NetworkManager*> GetNetworkManagers() {
-    return (NetworkManager::network_managers_);
-  }
+  NetworkThread *thread;          // reference to the libevent thread
+  ProtocolHandler protocol_handler_;       // Stores state for this socket
+  std::unique_ptr<PacketManager> packet_manager_;
+  ConnState state = CONN_INVALID;  // Initial state of connection
+  std::queue<std::unique_ptr<InputPacket>> rpkts_;                // Used for reading a single Postgres packet
 
   Client client_;
-
-  // has the startup packet been received for this connection
-  bool is_started = false;
-  bool ssl_sent = false;
-
-  // Should we send the buffered packets right away?
-  bool force_flush = false;
-
-  // TODO declare a response buffer pool so that we can reuse the responses
-  // so that we don't have to new packet each time
-  ResponseBuffer responses;
+ private:
+  Buffer rbuf_;                     // Socket's read buffer
+  Buffer wbuf_;                     // Socket's write buffer
+  unsigned int next_response_ = 0;  // The next response in the response buffer
 
  private:
-  //===--------------------------------------------------------------------===//
-  // PROTOCOL HANDLING FUNCTIONS
-  //===--------------------------------------------------------------------===//
+  // Is the requested amount of data available from the current position in
+  // the reader buffer?
+  bool IsReadDataAvailable(size_t bytes);
 
-  // Generic error protocol packet
-  void SendErrorResponse(
-      std::vector<std::pair<NetworkMessageType, std::string>> error_status);
+  // Parses out packet size from its header
+  void GetSizeFromPktHeader(size_t start_index);
 
-  // Sends ready for query packet to the frontend
-  void SendReadyForQuery(NetworkTransactionStateType txn_status);
-
-  // Sends the attribute headers required by SELECT queries
-  void PutTupleDescriptor(const std::vector<FieldInfo>& tuple_descriptor);
-
-  // Send each row, one packet at a time, used by SELECT queries
-  void SendDataRows(std::vector<StatementResult>& results, int colcount,
-                    int& rows_affected);
-
-  // Used to send a packet that indicates the completion of a query. Also has
-  // txn state mgmt
-  void CompleteCommand(const std::string& query_type_string, const QueryType& query_type, int rows);
-
-  // Specific response for empty or NULL queries
-  void SendEmptyQueryResponse();
-
-  /* Helper function used to make hardcoded ParameterStatus('S')
-   * packets during startup
+ public:
+  inline NetworkManager(int sock_fd, short event_flags, NetworkThread *thread,
+                        ConnState init_state)
+      : sock_fd(sock_fd) {
+    Init(event_flags, thread, init_state);
+  }
+  /** Get the initial packet from buffer initialize protocol handler
+   *
+   * @return true if we find the initial packet and
    */
-  void MakeHardcodedParameterStatus(
-      const std::pair<std::string, std::string>& kv);
+  bool GetInitialPacketFromBuffer();
 
-  /* We don't support "SET" and "SHOW" SQL commands yet.
-   * Also, duplicate BEGINs and COMMITs shouldn't be executed.
-   * This function helps filtering out the execution for such cases
+  /* Routine to deal with the first packet from the client */
+  int ProcessInitialPacket();
+
+  /* Routine to deal with general Startup message */
+  bool ProcessStartupPacket(std::string& contents, int32_t proto_version);
+
+  /* Routine to deal with SSL request message */
+  bool ProcessSSLRequestPacket();
+
+  /* Reuse this object for a new connection. We could be assigned to a
+   * new thread, change thread reference.
    */
-  bool HardcodedExecuteFilter(QueryType query_type);
+  void Init(short event_flags, NetworkThread *thread, ConnState init_state);
 
-  /* Execute a Simple query protocol message */
-  void ExecQueryMessage(InputPacket* pkt, const size_t thread_id);
+  /* refill_read_buffer - Used to repopulate read buffer with a fresh
+   * batch of data from the socket
+   */
+  ReadState FillReadBuffer();
 
-  /* Process the PARSE message of the extended query protocol */
-  void ExecParseMessage(InputPacket* pkt);
+  // Transit to the target state
+  void TransitState(ConnState next_state);
 
-  /* Process the BIND message of the extended query protocol */
-  void ExecBindMessage(InputPacket* pkt);
+  // Update the existing event to listen to the passed flags
+  bool UpdateEvent(short flags);
 
-  /* Process the DESCRIBE message of the extended query protocol */
-  bool ExecDescribeMessage(InputPacket* pkt);
+  WriteState WritePackets();
 
-  /* Process the EXECUTE message of the extended query protocol */
-  void ExecExecuteMessage(InputPacket* pkt, const size_t thread_id);
+  void PrintWriteBuffer();
 
-  /* Process the optional CLOSE message of the extended query protocol */
-  void ExecCloseMessage(InputPacket* pkt);
+  void CloseSocket();
 
-  //===--------------------------------------------------------------------===//
-  // MEMBERS
-  //===--------------------------------------------------------------------===//
+  void Reset();
 
-  // Manage standalone queries
-  std::shared_ptr<Statement> unnamed_statement_;
+ private:
+  // Writes a packet's header (type, size) into the write buffer
+  WriteState BufferWriteBytesHeader(OutputPacket *pkt);
 
-  // The result-column format code
-  std::vector<int> result_format_;
+  // Writes a packet's content into the write buffer
+  WriteState BufferWriteBytesContent(OutputPacket *pkt);
 
-  // global txn state
-  NetworkTransactionStateType txn_state_;
-
-  // state to mang skipped queries
-  bool skipped_stmt_ = false;
-  std::string skipped_query_string_;
-  QueryType skipped_query_type_;
-
-  // Statement cache
-  // StatementName -> Statement
-  Cache<std::string, Statement> statement_cache_;
-  // TableOid -> Statements
-  // FIXME: This table statement cache is not in sync with the other cache.
-  // That means if something gets thrown out of the statement cache it is not
-  // automatically evicted from this cache.
-  std::unordered_map<oid_t, std::vector<Statement*>> table_statement_cache_;
-
-  //  Portals
-  std::unordered_map<std::string, std::shared_ptr<Portal>> portals_;
-
-  // packets ready for read
-  size_t pkt_cntr_;
-
-  // Manage parameter types for unnamed statement
-  stats::QueryMetric::QueryParamBuf unnamed_stmt_param_types_;
-
-  // Parameter types for statements
-  // Warning: the data in the param buffer becomes invalid when the value
-  // stored
-  // in stat table is destroyed
-  std::unordered_map<std::string, stats::QueryMetric::QueryParamBuf>
-      statement_param_types_;
-
-  std::unique_ptr<PacketManger> packet_manager_;
-
-  // The traffic cop used for this connection
-  std::unique_ptr<tcop::TrafficCop> traffic_cop_;
-
-  //===--------------------------------------------------------------------===//
-  // STATIC DATA
-  //===--------------------------------------------------------------------===//
-
-  static const std::unordered_map<std::string, std::string>
-      parameter_status_map_;
-
-  // HACK: Global list of NetworkManager instances
-  // We need this in order to reset statement caches when the catalog changes
-  // We need to think of a more elegant solution for this
-  static std::vector<NetworkManager*> network_managers_;
-  static std::mutex packet_managers_mutex_;
-
+  // Used to invoke a write into the Socket, returns false if the socket is not
+  // ready for write
+  WriteState FlushWriteBuffer();
 };
 
-}  // End networking namespace
-}  // End peloton namespace
+}
+}
